requirements.txt
# Kaggle / pip install — minimal requirements
# Usage (Kaggle notebook):
#   ! pip install -r requirements.txt -q
#
# Note: torch + torchvision are pre-installed on Kaggle GPU images.

segmentation-models-pytorch>=0.3
albumentations>=1.3,<2.0
wandb
PyYAML
tqdm
scikit-learn
matplotlib
Pillow
numpy
